fixed category & refactor the code structure and build

+++ src/Deed/Deed.Application/Capitals/Commands/Create/CreateCapitalCommand.cs
@@ -5,7 +5,7 @@ namespace Deed.Application.Capitals.Commands.Create;
 
 public sealed record CreateCapitalCommand(
     string Name,
-    float Balance,
+    decimal Balance,
     CurrencyType Currency,
     bool IncludeInTotal,
     bool OnlyForSavings)
