+++ src/Deed/Deed.Application/Capitals/Requests/CreateCapitalRequest.cs
@@ -4,7 +4,7 @@ namespace Deed.Application.Capitals.Requests;
 
 public sealed record CreateCapitalRequest(
     string Name,
-    float Balance,
+    decimal Balance,
     CurrencyType Currency,
     bool IncludeInTotal,
     bool OnlyForSavings
