+++ src/Deed/Deed.Application/Expenses/Commands/Update/UpdateExpenseCommand.cs
@@ -5,7 +5,7 @@ namespace Deed.Application.Expenses.Commands.Update;
 public sealed record UpdateExpenseCommand(
     int Id,
     int? CategoryId = null,
-    float? Amount = null,
+    decimal? Amount = null,
     string? Purpose = null,
     DateTimeOffset? Date = null)
     : ICommand;
