+++ src/Deed/Deed.Application/Expenses/Requests/CreateExpenseRequest.cs
@@ -3,6 +3,6 @@
 public sealed record CreateExpenseRequest(
     int CapitalId,
     int CategoryId,
-    float Amount,
+    decimal Amount,
     DateTimeOffset PaymentDate,
     string? Purpose);
