+++ src/Deed/Deed.Application/Incomes/Responses/IncomeResponse.cs
@@ -3,7 +3,7 @@ namespace Deed.Application.Incomes.Responses;
 public sealed record IncomeResponse(
     int Id,
     int CategoryId,
-    float Amount,
+    decimal Amount,
     string? Purpose,
     DateTimeOffset CreatedAt,
     int? CapitalId);
