+++ tests/Deed.Tests.Unit/Categories/Commands/DeleteCategoryCommandHandlerTests.cs
@@ -22,12 +22,12 @@ public sealed class DeleteCategoryCommandHandlerTests
     }
 
     [Theory]
-    [InlineData(CategoryType.Incomes, "IncomesWithHighAmountOfDollars", 0f, PerPeriodType.None)]
-    [InlineData(CategoryType.Expenses, "E", 100f, PerPeriodType.Daily)]
+    [InlineData(CategoryType.Incomes, "IncomesWithHighAmountOfDollars", 0, PerPeriodType.None)]
+    [InlineData(CategoryType.Expenses, "E", 100, PerPeriodType.Daily)]
     public async Task Handle_ShouldDeleteCategory_ReturnSuccess(
         CategoryType type,
         string name,
-        float periodAmount,
+        decimal periodAmount,
         PerPeriodType periodType)
     {
         // Arrange
